Updated job feed to not display jobs created by the user browsing the job feed (#118)
Updated the function `get_all(self)` to `get_all(self, owner_id=None)`
in ***class JobView*** in views.py. If the user is an owner and a
sitter, the call to `get_all` is passed **request.user.id** as the
owner_id, which excludes jobs created by that user and returns all other
jobs.