Fixed date/time bug and default location behavior bug (#129)
1. Initially, the time of the job added by the owner and the job
displayed in the jobs list of the owner and the job feed of the sitter
were different due to a discrepancy in time zone. Now, the date and time
represented are of the same time zone.
2. Initially, when a job has been created and saved, the fields of the
create job form are emptied. However, we want the location to revert
back to the default location instead of empty. That has been fixed.
3. Added location delete functionality.

![Screenshot 2023-12-11
073606](https://github.com/gcivil-nyu-org/INET-Monday-Fall2023-Team-1/assets/71821708/bbf265b9-c508-4c53-88b4-a307a3104934)

---------

Co-authored-by: Sameer Kolhar <kolhar730@gmail.com>